--- app/assets/javascripts/admin/addon/components/  modal/delete-posts-confirmation.hbs
@@ -1,19 +1,0 @@
-<DModal @title={{html-safe this.title}} @closeModal={{@closeModal}}>
-  <:body>
-    <p>{{html-safe this.description}}</p>
-    <Input @type="text" @value={{this.value}} />
-  </:body>
-  <:footer>
-    <DButton
-      class="btn-danger"
-      @action={{this.confirm}}
-      @icon="trash-alt"
-      @disabled={{this.deleteDisabled}}
-      @translatedLabel={{this.deleteButtonText}}
-    />
-    <DButton
-      @action={{@closeModal}}
-      @label="admin.user.delete_posts.confirmation.cancel"
-    />
-  </:footer>
-</DModal>
