+++ app/assets/javascripts/discourse/app/services/modal.js
@@ -55,7 +55,6 @@ const KNOWN_LEGACY_MODALS = [
   "topic-summary",
   "user-status",
   "admin-add-upload",
-  "admin-delete-posts-confirmation",
   "admin-merge-users-prompt",
   "admin-start-backup",
   "admin-watched-word-test",
